requirements.txt
Babel==2.8.0
blinker==1.4
click==7.1.2
dnspython==1.16.0
email_validator==1.1.1
flask==1.1.2
flask_cors==3.0.8
flask_httpauth==4.0.0
flask_login==0.5.0
flask_mail==0.9.1
flask_mongoengine==0.9.5
flask_security==3.0.0
flask_wtf==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
passlib==1.7.2
pymongo==3.10.1
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
speaklater==1.3
Werkzeug==1.0.1
WTForms==2.3.1
